+++ installer/templates/phx_umbrella/apps/app_name_web/lib/app_name.ex
@@ -34,7 +34,8 @@ defmodule <%= web_namespace %> do
         namespace: <%= web_namespace %>
 
       # Import convenience functions from controllers
-      import Phoenix.Controller, only: [get_flash: 1, get_flash: 2, view_module: 1]
+      import Phoenix.Controller,
+        only: [get_flash: 1, get_flash: 2, view_module: 1, view_template: 1]
 
       # Include shared imports and aliases for views
       unquote(view_helpers())