+++ lib/phoenix/view.ex
@@ -20,7 +20,8 @@ defmodule Phoenix.View do
             use Phoenix.View, root: "lib/your_app_web/templates", namespace: "web"
 
             # Import convenience functions from controllers
-            import Phoenix.Controller, only: [get_flash: 1, get_flash: 2, view_module: 1]
+            import Phoenix.Controller,
+              only: [get_flash: 1, get_flash: 2, view_module: 1, view_template: 1]
 
             # Use all HTML functionality (forms, tags, etc)
             use Phoenix.HTML
@@ -259,11 +260,6 @@ defmodule Phoenix.View do
     * `:layout` - tells Phoenix to wrap the rendered result in the
       given layout. See next section
 
-  The following assigns are reserved, and cannot be set directly:
-
-    * `@view_module` - The view module being rendered
-    * `@view_template` - The `@view_module`'s template being rendered
-
   ## Layouts
 
   Templates can be rendered within other templates using the `:layout`
@@ -278,25 +274,9 @@ defmodule Phoenix.View do
   """
   def render(module, template, assigns)
 
-  def render(module, template, %{deprecated_module_template: {module, template, content}}) do
-    IO.warn """
-    Rendering the child template from layouts is deprecated. Instead of:
-
-        <%= render(@view_module, @view_template, assigns) %>
-
-    You should do:
-
-        <%= @inner_content %>
-    """
-
-    content
-  end
-
   def render(module, template, assigns) do
     assigns
     |> Map.new()
-    |> Map.put(:view_module, module)
-    |> Map.put(:view_template, template)
     |> Map.pop(:layout, false)
     |> render_within(module, template)
   end
@@ -305,23 +285,14 @@ defmodule Phoenix.View do
     module.render(template, assigns)
   end
 
-  defp render_within({layout, assigns}, module, template) do
-    content = module.render(template, assigns)
-
-    assigns =
-      assigns
-      |> Map.put(:inner_content, content)
-      |> Map.put(:deprecated_module_template, {module, template, content})
-
-    render_layout(layout, assigns)
-  end
-
-  defp render_layout({layout_mod, layout_tpl}, assigns)
+  defp render_within({{layout_mod, layout_tpl}, assigns}, module, template)
        when is_atom(layout_mod) and is_binary(layout_tpl) do
+    content = module.render(template, assigns)
+    assigns = Map.put(assigns, :inner_content, content)
     layout_mod.render(layout_tpl, assigns)
   end
 
-  defp render_layout(layout, _assigns) do
+  defp render_within({layout, _assigns}, _module, _template) do
     raise ArgumentError, """
     invalid value for reserved key :layout in View.render/3 assigns
 
@@ -335,8 +306,9 @@ defmodule Phoenix.View do
   Renders a template only if it exists.
 
   Same as `render/3`, but returns `nil` instead of raising.
-  Useful for dynamically rendering templates in the layout that may or
-  may not be implemented by the `@view_module` view.
+  This is often used with `Phoenix.Controller.view_module/1`
+  and `Phoenix.Controller.view_template/1`, which must be
+  imported into your views. See the "Examples" section below.
 
   ## Examples
 
@@ -345,10 +317,10 @@ defmodule Phoenix.View do
   may wish to inject certain scripts, while others will not.
 
       <head>
-        <%= render_existing @view_module, "scripts.html", assigns %>
+        <%= render_existing view_module(@conn), "scripts.html", assigns %>
       </head>
 
-  Then the module for the `@view_module` view can decide to provide scripts with
+  Then the module under `view_module(@conn)` can decide to provide scripts with
   either a precompiled template, or by implementing the function directly, ie:
 
       def render("scripts.html", _assigns) do
@@ -366,7 +338,7 @@ defmodule Phoenix.View do
   `render_existing/3` for per-template based content, ie:
 
       <head>
-        <%= render_existing @view_module, "scripts." <> @view_template, assigns %>
+        <%= render_existing view_module(@conn), "scripts." <> view_template(@conn), assigns %>
       </head>
 
       def render("scripts.show.html", _assigns) do