+++ test/fixtures/views.exs
@@ -32,6 +32,8 @@ end
 defmodule MyApp.UserView do
   use Phoenix.View, root: "test/fixtures/templates", pattern: "**/*"
 
+  import Phoenix.Controller, only: [view_module: 1, view_template: 1]
+
   def escaped_title(title) do
     {:safe, Plug.HTML.html_escape(title)}
   end
@@ -61,7 +63,7 @@ defmodule MyApp.UserView do
 
   def render("inner.html", assigns) do
     """
-    View module is #{assigns.view_module} and view template is #{assigns.view_template}
+    View module is #{view_module(assigns.conn)} and view template is #{view_template(assigns.conn)}
     """
   end